+++ app/src/main/java/de/baumann/browser/popups/Popup_readLater.java
@@ -71,8 +71,16 @@ public class Popup_readLater extends Activity {
         button.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Popup_readLater.this.deleteDatabase("readLater.db");
-                setBookmarkList();
+                Snackbar snackbar = Snackbar
+                        .make(listView, R.string.toast_list, Snackbar.LENGTH_LONG)
+                        .setAction(R.string.toast_yes, new View.OnClickListener() {
+                            @Override
+                            public void onClick(View view) {
+                                Popup_readLater.this.deleteDatabase("readLater.db");
+                                setBookmarkList();
+                            }
+                        });
+                snackbar.show();
             }
         });
         ImageButton buttonSort = (ImageButton) findViewById(R.id.butSort);